fix 6-size,c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float e;
-	printf("Size of a char: %ld byte(s)", sizeof(a));
-	Printf("Size of an int: %d byte(s)", sizeof(b));
-	printf("Size of a long int: %ld byte(s)", sizeof(c));
-	printf("Size of a long long int: %d byte(s)", sizeof(d));
-	printf("Size of a float: %d byte(s)", sizeof(e));
+	printf("Size of a char: %d byte(s)", (int)sizeof(a));
+	Printf("Size of an int: %d byte(s)", (int)sizeof(b));
+	printf("Size of a long int: %d byte(s)", (int)sizeof(c));
+	printf("Size of a long long int: %d byte(s)", (int)sizeof(d));
+	printf("Size of a float: %d byte(s)", (int)sizeof(e));
 	return (0);
 }
